src/telemetry_forager/requirements.txt
asgiref==3.8.1
azure-common==1.1.28
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-functions==1.23.0
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.3.0
azure-mgmt-resourcehealth==1.0.0b6
azure-monitor-opentelemetry==1.6.7
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-monitor-query==1.4.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
fixedint==0.1.6
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
isodate==0.7.2
jsons==1.6.3
MarkupSafe==3.0.2
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
numpy==2.2.4
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
opentelemetry-api==1.32.0
opentelemetry-instrumentation==0.53b0
opentelemetry-instrumentation-asgi==0.53b0
opentelemetry-instrumentation-dbapi==0.53b0
opentelemetry-instrumentation-django==0.53b0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-instrumentation-flask==0.53b0
opentelemetry-instrumentation-psycopg2==0.53b0
opentelemetry-instrumentation-requests==0.53b0
opentelemetry-instrumentation-urllib==0.53b0
opentelemetry-instrumentation-urllib3==0.53b0
opentelemetry-instrumentation-wsgi==0.53b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
opentelemetry-util-http==0.53b0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
typing_extensions==4.13.2
typish==1.9.3
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
